Extract common functionality between  and  into a new location to prevent code repetition
