+++ language/go/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "resolve.go",
         "std_package_list.go",
         "update.go",
+        "utils.go",
         "work.go",
     ],
     importpath = "github.com/bazelbuild/bazel-gazelle/language/go",
@@ -113,6 +114,7 @@ filegroup(
         "stubs_test.go",
         "update.go",
         "update_import_test.go",
+        "utils.go",
         "work.go",
         "//language/go/gen_std_package_list:all_files",
     ],
